+++ README.md
@@ -54,21 +54,14 @@
 
 ### TorBot
 
-#### Using `poetry`
-* TorBot dependencies are managed using `poetry`, you can find the installation commands below:
-```sh
-poetry install # to install dependencies
-poetry run python torbot -u https://www.example.com --depth 2 --visualize tree --save json # example of running command with poetry
-poetry run python torbot -h # for help
-```
-
 #### Using `venv`
 * If using Python ^3.4,
 ```sh
 python -m venv torbot_venv
 source torbot_venv/bin/activate
 pip install -r requirements.txt
-python torbot -u https://www.example.com
+pip install -e .
+./main.py --help
 ```
 
 #### Using `docker`
