+++ src/torbot/modules/color.py
